need to cast this to integer or it is taken as a name, not id.
yea unit tests

+++ content/lib/Tagger.php
@@ -808,7 +808,7 @@ class Content_Tagger
     {
         if (is_array($object)) {
             $object = current($this->_objectManager->ensureObjects(
-                $object['object'], current($this->_typeManager->ensureTypes($object['type']))));
+                $object['object'], (int)current($this->_typeManager->ensureTypes($object['type']))));
         }
 
         return (int)$object;
